Create a variable for empty spot in Game
Prior to this change, the empty spot on the board was represented by the
space character ' '. This meant that everywhere where the empty spot was
used, it was used directly. This made changing the character hard.

This change turns this character into a variable. This way, if we want
to change the representation of the empty spot on the board, we simply
have to change one variable.