Bugfix invalid move generation
Prior to this change, the private helper method canJump(int, int,
ColumnDirection) gave invalid checks to the move generator, because it
did not check whether there was an enemy in the jump direction.

This change makes canJump(int, int, ColumnDirection) check whether an
enemy is present between the jump tiles. This corrects move generation.