Create an extremely basic GUI
Prior to this change, this app was entirely CLI based. The project
requirements needed me to create a JavaFX interface.

This change creates a grid layout of buttons which change text based on
the board state. The player can move the pieces as before. There is no
computer mode nor are there any labels or warnings when an exception
gets thrown.